Performance Improvement
The performance of the build was slow because it always had to install composer and npm packages.
Overall performance improvement.

Changed the Dockerfile to make more use of layer caching by adding the code later.
Added caching to the volumes for better performance overall.
Added chrome to app.catroweb.dev container for more stable testing.