SHARE-162 Refactoring tests (behat)
To improve testing with docker and make it more convenient to work with, behat and parameters file was replaced with a regex in the dockerfile.
Chrome container had no access to DataFixture folder.
Also fixed some tests that did not work.

Changed Dockerfile
Added volume of DataFixture to Chrome container
Changed feature files to work even on smaller systems